refactor: modularize captain task distribution
