--- scripts/captain_task_distribution_simple.py
@@ -1,666 +1,0 @@
-#!/usr/bin/env python3
-"""
-Captain Task Distribution Script - Agent Cellphone V2 (Simplified)
-================================================================
-
-Distributes all 13 remaining tasks to Agents 1-4 with comprehensive coding standards
-and expected outcomes included in each contract.
-
-Author: Agent-4 (Captain)
-Purpose: Final task distribution for Phase 4 completion
-"""
-
-import json
-import time
-import logging
-from pathlib import Path
-from datetime import datetime
-
-# Configure logging
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
-
-class CaptainTaskDistributor:
-    """Captain Agent-4 Task Distribution System (Simplified)"""
-    
-    def __init__(self):
-        """Initialize the task distribution system"""
-        self.tasks_distributed = 0
-        self.output_dir = Path("logs")
-        self.output_dir.mkdir(exist_ok=True)
-        
-        logger.info("🚀 Captain Task Distributor initialized")
-    
-    def get_coding_standards_contract(self) -> str:
-        """Get comprehensive coding standards for all agents"""
-        return """
-🚨 **CRITICAL CODING STANDARDS - AGENT CELLPHONE V2**
-
-**🎯 PRIMARY PRINCIPLE: USE EXISTING ARCHITECTURE FIRST**
-- **NEVER** create duplicate functionality
-- **NEVER** build new systems when existing ones work
-- **ALWAYS** extend existing architecture first
-- **ALWAYS** integrate with current systems
-
-**🏗️ ARCHITECTURE PRINCIPLES**
-1. **Single Responsibility Principle (SRP)**: Each class/module has ONE clear purpose
-2. **Open/Closed Principle (OCP)**: Open for extension, closed for modification
-3. **Dependency Inversion Principle (DIP)**: Depend on existing abstractions
-
-**🔧 IMPLEMENTATION REQUIREMENTS**
-- **Check existing systems first** before any new implementation
-- **Extend existing classes** rather than create new ones
-- **Use existing interfaces** and contracts
-- **Integrate with current infrastructure**
-- **Follow V2 standards**: ≤200 LOC per file, OOP design, SRP compliance
-
-**📱 EXISTING SYSTEMS TO USE**
-- Coordinate Manager: `src/services/messaging/coordinate_manager.py`
-- PyAutoGUI Messaging: `src/services/messaging/unified_pyautogui_messaging.py`
-- Agent Coordinator: `src/autonomous_development/agents/agent_coordinator.py`
-- Message Coordinator: `src/services/communication/message_coordinator.py`
-- Task Scheduler: `src/services/communication/task_scheduler_coordinator.py`
-- Workflow Engine: `src/services/communication/workflow_engine.py`
-
-**❌ WHAT NOT TO DO**
-- Don't create new messaging systems
-- Don't duplicate existing functionality
-- Don't ignore existing architecture
-- Don't create files >200 LOC
-- Don't violate SRP principles
-
-**✅ WHAT TO DO**
-- Extend existing systems with new methods
-- Use existing interfaces and contracts
-- Leverage existing coordinate systems
-- Integrate with current messaging infrastructure
-- Build on existing agent management
-"""
-
-    def create_agent_1_contracts(self) -> list:
-        """Create comprehensive contracts for Agent-1 tasks"""
-        contracts = []
-        
-        # TASK 1H - Phase 1 Finalization
-        contract_1h = f"""
-🎯 **AGENT-1 CONTRACT: TASK 1H - PHASE 1 FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Complete Phase 1 consolidation by finalizing remaining 3 systems integration
-
-**📋 DELIVERABLES**:
-1) Complete Phase 1 system consolidation (3 remaining systems)
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document all consolidation tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate no duplicate functionality exists
-
-**✅ EXPECTED RESULTS**:
-- Phase 1 consolidation 100% complete
-- All 3 remaining systems integrated with existing architecture
-- Devlog created with detailed completion documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- No duplicate functionality introduced
-- All systems use existing unified infrastructure
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-1", contract_1h, "TASK 1H"))
-        
-        # TASK 1I - Workflow Integration Finalization
-        contract_1i = f"""
-🎯 **AGENT-1 CONTRACT: TASK 1I - WORKFLOW INTEGRATION FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize workflow system integration using existing unified systems
-
-**📋 DELIVERABLES**:
-1) Complete workflow engine integration with existing systems
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document integration tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate workflow system uses existing messaging infrastructure
-
-**✅ EXPECTED RESULTS**:
-- Workflow engine fully integrated with existing systems
-- Devlog created with detailed integration documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Workflow system uses existing unified messaging
-- No duplicate workflow functionality
-- All integration follows existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-1", contract_1i, "TASK 1I"))
-        
-        # TASK 1J - Workflow Engine Finalization
-        contract_1j = f"""
-🎯 **AGENT-1 CONTRACT: TASK 1J - WORKFLOW ENGINE FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize workflow engine using existing BaseManager patterns
-
-**📋 DELIVERABLES**:
-1) Complete workflow engine finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document engine finalization tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate workflow engine follows existing patterns
-
-**✅ EXPECTED RESULTS**:
-- Workflow engine 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Engine follows existing BaseManager patterns
-- No duplicate engine functionality
-- All workflows use existing unified systems
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-1", contract_1j, "TASK 1J"))
-        
-        # TASK 1K - Learning System Finalization
-        contract_1k = f"""
-🎯 **AGENT-1 CONTRACT: TASK 1K - LEARNING SYSTEM FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize learning system integration using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete learning system finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document learning system tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate learning system uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Learning system 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- System uses existing unified infrastructure
-- No duplicate learning functionality
-- All learning follows existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-1", contract_1k, "TASK 1K"))
-        
-        return contracts
-    
-    def create_agent_2_contracts(self) -> list:
-        """Create comprehensive contracts for Agent-2 tasks"""
-        contracts = []
-        
-        # TASK 2H - Manager System Finalization
-        contract_2h = f"""
-🎯 **AGENT-2 CONTRACT: TASK 2H - MANAGER SYSTEM FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize manager system using existing BaseManager patterns
-
-**📋 DELIVERABLES**:
-1) Complete manager system finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document manager system tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate all managers follow existing BaseManager patterns
-
-**✅ EXPECTED RESULTS**:
-- Manager system 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- All managers extend existing BaseManager
-- No duplicate manager functionality
-- All managers use existing unified infrastructure
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-2", contract_2h, "TASK 2H"))
-        
-        # TASK 2I - Performance System Finalization
-        contract_2i = f"""
-🎯 **AGENT-2 CONTRACT: TASK 2I - PERFORMANCE SYSTEM FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize performance system using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete performance system finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document performance system tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate performance system uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Performance system 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- System uses existing unified infrastructure
-- No duplicate performance functionality
-- All performance follows existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-2", contract_2i, "TASK 2I"))
-        
-        # TASK 2J - Health System Finalization
-        contract_2j = f"""
-🎯 **AGENT-2 CONTRACT: TASK 2J - HEALTH SYSTEM FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize health system using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete health system finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document health system tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate health system uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Health system 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- System uses existing unified infrastructure
-- No duplicate health functionality
-- All health monitoring follows existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-2", contract_2j, "TASK 2J"))
-        
-        # TASK 2K - API Integration Finalization
-        contract_2k = f"""
-🎯 **AGENT-2 CONTRACT: TASK 2K - API INTEGRATION FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize API integration using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete API integration finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document API integration tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate API integration uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- API integration 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Integration uses existing unified infrastructure
-- No duplicate API functionality
-- All API calls follow existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-2", contract_2k, "TASK 2K"))
-        
-        return contracts
-    
-    def create_agent_3_contracts(self) -> list:
-        """Create comprehensive contracts for Agent-3 tasks"""
-        contracts = []
-        
-        # TASK 3G - Testing Infrastructure Finalization
-        contract_3g = f"""
-🎯 **AGENT-3 CONTRACT: TASK 3G - TESTING INFRASTRUCTURE FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize testing infrastructure using existing unified systems
-
-**📋 DELIVERABLES**:
-1) Complete testing infrastructure finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document testing infrastructure tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate testing uses existing unified infrastructure
-
-**✅ EXPECTED RESULTS**:
-- Testing infrastructure 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Infrastructure uses existing unified systems
-- No duplicate testing functionality
-- All tests follow existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-3", contract_3g, "TASK 3G"))
-        
-        # TASK 3H - Reporting Systems Finalization
-        contract_3h = f"""
-🎯 **AGENT-3 CONTRACT: TASK 3H - REPORTING SYSTEMS FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize reporting systems using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete reporting systems finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document reporting systems tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate reporting uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Reporting systems 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Systems use existing unified infrastructure
-- No duplicate reporting functionality
-- All reporting follows existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-3", contract_3h, "TASK 3H"))
-        
-        # TASK 3I - Integration Testing Finalization
-        contract_3i = f"""
-🎯 **AGENT-3 CONTRACT: TASK 3I - INTEGRATION TESTING FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize integration testing using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete integration testing finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document integration testing tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate integration testing uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Integration testing 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Testing uses existing unified infrastructure
-- No duplicate testing functionality
-- All integration tests follow existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-3", contract_3i, "TASK 3I"))
-        
-        # TASK 3J - Model Consolidation Finalization
-        contract_3j = f"""
-🎯 **AGENT-3 CONTRACT: TASK 3J - MODEL CONSOLIDATION FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize model consolidation using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete model consolidation finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document model consolidation tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate model consolidation uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Model consolidation 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- Consolidation uses existing unified infrastructure
-- No duplicate model functionality
-- All models follow existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-3", contract_3j, "TASK 3J"))
-        
-        return contracts
-    
-    def create_agent_4_contracts(self) -> list:
-        """Create comprehensive contracts for Agent-4 tasks"""
-        contracts = []
-        
-        # TASK 4J - Repository System Finalization
-        contract_4j = f"""
-🎯 **AGENT-4 CONTRACT: TASK 4J - REPOSITORY SYSTEM FINALIZATION**
-
-{self.get_coding_standards_contract()}
-
-**🎯 OBJECTIVE**: Finalize repository system using existing unified infrastructure
-
-**📋 DELIVERABLES**:
-1) Complete repository system finalization
-2) Create comprehensive devlog entry in `logs/` directory
-3) Document repository system tasks completed
-4) Report architecture compliance status (V2 standards)
-5) Validate repository system uses existing unified systems
-
-**✅ EXPECTED RESULTS**:
-- Repository system 100% complete and functional
-- Devlog created with detailed finalization documentation
-- Architecture compliance verified (≤200 LOC, SRP, OOP)
-- System uses existing unified infrastructure
-- No duplicate repository functionality
-- All repository operations follow existing patterns
-
-**⏰ TIMELINE**: 2-3 hours
-
-**🚀 START NOW. WE. ARE. SWARM.**
-"""
-        contracts.append(("Agent-4", contract_4j, "TASK 4J"))
-        
-        return contracts
-    
-    def save_contracts_to_files(self) -> bool:
-        """Save all contracts to individual files for distribution"""
-        logger.info("🚀 Starting comprehensive task contract creation")
-        
-        # Create all contracts
-        all_contracts = []
-        all_contracts.extend(self.create_agent_1_contracts())
-        all_contracts.extend(self.create_agent_2_contracts())
-        all_contracts.extend(self.create_agent_3_contracts())
-        all_contracts.extend(self.create_agent_4_contracts())
-        
-        logger.info(f"📋 Created {len(all_contracts)} comprehensive task contracts")
-        
-        # Save contracts to files
-        success_count = 0
-        for agent, contract_content, task_id in all_contracts:
-            try:
-                # Create agent directory
-                agent_dir = self.output_dir / f"contracts_{agent.lower().replace('-', '_')}"
-                agent_dir.mkdir(exist_ok=True)
-                
-                # Save contract file
-                contract_file = agent_dir / f"{task_id.lower().replace(' ', '_')}.md"
-                with open(contract_file, 'w', encoding='utf-8') as f:
-                    f.write(contract_content)
-                
-                success_count += 1
-                logger.info(f"✅ {task_id} contract saved for {agent}: {contract_file}")
-                self.tasks_distributed += 1
-                
-            except Exception as e:
-                logger.error(f"❌ Error saving {task_id} contract for {agent}: {e}")
-        
-        # Create master contract summary
-        self._create_master_summary(all_contracts)
-        
-        logger.info(f"🎯 Contract creation complete: {success_count}/{len(all_contracts)} successful")
-        return success_count == len(all_contracts)
-    
-    def _create_master_summary(self, all_contracts: list) -> None:
-        """Create master summary of all distributed contracts"""
-        summary_file = self.output_dir / "CAPTAIN_TASK_DISTRIBUTION_SUMMARY.md"
-        
-        summary_content = f"""# 🎯 CAPTAIN TASK DISTRIBUTION SUMMARY - AGENT CELLPHONE V2
-
-**Captain**: Agent-4  
-**Distribution Date**: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}  
-**Total Tasks Distributed**: {len(all_contracts)}  
-**Agents Covered**: Agent-1, Agent-2, Agent-3, Agent-4
-
----
-
-## 📊 **TASK DISTRIBUTION OVERVIEW**
-
-### **AGENT-1 TASKS (4 tasks)**
-- **TASK 1H**: Phase 1 Finalization
-- **TASK 1I**: Workflow Integration Finalization  
-- **TASK 1J**: Workflow Engine Finalization
-- **TASK 1K**: Learning System Finalization
-
-### **AGENT-2 TASKS (4 tasks)**
-- **TASK 2H**: Manager System Finalization
-- **TASK 2I**: Performance System Finalization
-- **TASK 2J**: Health System Finalization
-- **TASK 2K**: API Integration Finalization
-
-### **AGENT-3 TASKS (4 tasks)**
-- **TASK 3G**: Testing Infrastructure Finalization
-- **TASK 3H**: Reporting Systems Finalization
-- **TASK 3I**: Integration Testing Finalization
-- **TASK 3J**: Model Consolidation Finalization
-
-### **AGENT-4 TASKS (1 task)**
-- **TASK 4J**: Repository System Finalization
-
----
-
-## 🚨 **COMPREHENSIVE CODING STANDARDS INCLUDED**
-
-All contracts include:
-- ✅ **Primary Principle**: Use existing architecture first
-- ✅ **Architecture Principles**: SRP, OCP, DIP compliance
-- ✅ **Implementation Requirements**: V2 standards (≤200 LOC, OOP, SRP)
-- ✅ **Existing Systems**: Specific paths to use
-- ✅ **What NOT to do**: Clear violation examples
-- ✅ **What TO do**: Proper implementation patterns
-
----
-
-## 📋 **CONTRACT STRUCTURE STANDARD**
-
-Each contract includes:
-- 🎯 **Clear Objective**: What needs to be accomplished
-- 📋 **Specific Deliverables**: What must be produced
-- ✅ **Expected Results**: What success looks like
-- ⏰ **Timeline**: 2-3 hours per task
-- 🚀 **Action Command**: Clear instruction to begin
-
----
-
-## 🎯 **EXPECTED OUTCOMES**
-
-### **Immediate (Next 2-3 hours)**
-- All 13 tasks distributed with comprehensive standards
-- Agents have clear, actionable contracts
-- Architecture compliance requirements specified
-- V2 coding standards integrated
-
-### **Short Term (Next 6-8 hours)**
-- All Phase 1-4 tasks completed
-- 100% task completion achieved
-- Full system unification complete
-- Architecture compliance verified
-
-### **Medium Term (Next 2-3 weeks)**
-- Technical debt reduction (80% target)
-- Large file modularization (44 files >200 LOC)
-- TODO implementation (20+ critical functions)
-- Debug standardization
-
----
-
-## 🚀 **MISSION STATUS**
-
-**WE. ARE. SWARM.** 🚀
-
-All 13 remaining tasks have been distributed with comprehensive coding standards and expected outcomes. Each agent now has clear, actionable contracts that specify exactly what V2 standards to follow and what deliverables to produce.
-
-**Next Phase**: Execute all tasks in parallel across Agents 1-4 to achieve 100% completion and full system unification.
-
----
-
-**Generated by**: Captain Agent-4 Task Distribution System  
-**Timestamp**: {datetime.now().isoformat()}
-"""
-        
-        with open(summary_file, 'w', encoding='utf-8') as f:
-            f.write(summary_content)
-        
-        logger.info(f"📊 Master summary created: {summary_file}")
-    
-    def get_distribution_summary(self) -> dict:
-        """Get summary of task distribution"""
-        return {
-            "total_tasks": 13,
-            "tasks_distributed": self.tasks_distributed,
-            "distribution_rate": f"{(self.tasks_distributed / 13) * 100:.1f}%",
-            "agents_covered": ["Agent-1", "Agent-2", "Agent-3", "Agent-4"],
-            "coding_standards_included": True,
-            "expected_outcomes_specified": True,
-            "architecture_compliance_required": True,
-            "output_directory": str(self.output_dir)
-        }
-
-def main():
-    """Main execution function"""
-    logger.info("🚀 Captain Agent-4 Task Distribution System Starting")
-    
-    try:
-        # Initialize distributor
-        distributor = CaptainTaskDistributor()
-        
-        # Create and save all contracts
-        success = distributor.save_contracts_to_files()
-        
-        if success:
-            logger.info("🎉 All task contracts successfully created with comprehensive standards!")
-        else:
-            logger.warning("⚠️ Some contracts failed to create - check logs")
-        
-        # Get summary
-        summary = distributor.get_distribution_summary()
-        logger.info(f"📊 Distribution Summary: {summary}")
-        
-        logger.info("🎯 Task distribution complete. All agents have comprehensive contracts!")
-        logger.info(f"📁 Contracts saved to: {distributor.output_dir}")
-        
-    except Exception as e:
-        logger.error(f"❌ Critical error in task distribution: {e}")
-        return False
-    
-    return True
-
-if __name__ == "__main__":
-    success = main()
-    exit(0 if success else 1)
