+++ src/task_distribution/agents/agent1.py
@@ -1,0 +1,122 @@
+from __future__ import annotations
+
+from ..models import TaskContract
+from ..standards import CODING_STANDARDS
+
+
+class Agent1Contracts:
+    """Generate contracts for Agent-1 tasks."""
+
+    @staticmethod
+    def build() -> list[TaskContract]:
+        contracts: list[TaskContract] = []
+
+        contract_1h = f"""🎯 **AGENT-1 CONTRACT: TASK 1H - PHASE 1 FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Complete Phase 1 consolidation by finalizing remaining 3 systems integration
+
+**📋 DELIVERABLES**:
+1) Complete Phase 1 system consolidation (3 remaining systems)
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document all consolidation tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate no duplicate functionality exists
+
+**✅ EXPECTED RESULTS**:
+- Phase 1 consolidation 100% complete
+- All 3 remaining systems integrated with existing architecture
+- Devlog created with detailed completion documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- No duplicate functionality introduced
+- All systems use existing unified infrastructure
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-1", "TASK 1H", contract_1h))
+
+        contract_1i = f"""🎯 **AGENT-1 CONTRACT: TASK 1I - WORKFLOW INTEGRATION FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize workflow system integration using existing unified systems
+
+**📋 DELIVERABLES**:
+1) Complete workflow engine integration with existing systems
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document integration tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate workflow system uses existing messaging infrastructure
+
+**✅ EXPECTED RESULTS**:
+- Workflow engine fully integrated with existing systems
+- Devlog created with detailed integration documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Workflow system uses existing unified messaging
+- No duplicate workflow functionality
+- All integration follows existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-1", "TASK 1I", contract_1i))
+
+        contract_1j = f"""🎯 **AGENT-1 CONTRACT: TASK 1J - WORKFLOW ENGINE FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize workflow engine using existing BaseManager patterns
+
+**📋 DELIVERABLES**:
+1) Complete workflow engine finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document engine finalization tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate workflow engine follows existing patterns
+
+**✅ EXPECTED RESULTS**:
+- Workflow engine 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Engine follows existing BaseManager patterns
+- No duplicate engine functionality
+- All workflows use existing unified systems
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-1", "TASK 1J", contract_1j))
+
+        contract_1k = f"""🎯 **AGENT-1 CONTRACT: TASK 1K - LEARNING SYSTEM FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize learning system integration using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete learning system finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document learning system tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate learning system uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Learning system 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- System uses existing unified infrastructure
+- No duplicate learning functionality
+- All learning follows existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-1", "TASK 1K", contract_1k))
+
+        return contracts
