+++ src/task_distribution/agents/agent2.py
@@ -1,0 +1,122 @@
+from __future__ import annotations
+
+from ..models import TaskContract
+from ..standards import CODING_STANDARDS
+
+
+class Agent2Contracts:
+    """Generate contracts for Agent-2 tasks."""
+
+    @staticmethod
+    def build() -> list[TaskContract]:
+        contracts: list[TaskContract] = []
+
+        contract_2h = f"""🎯 **AGENT-2 CONTRACT: TASK 2H - MANAGER SYSTEM FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize manager system using existing BaseManager patterns
+
+**📋 DELIVERABLES**:
+1) Complete manager system finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document manager system tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate all managers follow existing BaseManager patterns
+
+**✅ EXPECTED RESULTS**:
+- Manager system 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- All managers extend existing BaseManager
+- No duplicate manager functionality
+- All managers use existing unified infrastructure
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-2", "TASK 2H", contract_2h))
+
+        contract_2i = f"""🎯 **AGENT-2 CONTRACT: TASK 2I - PERFORMANCE SYSTEM FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize performance system using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete performance system finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document performance system tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate performance system uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Performance system 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- System uses existing unified infrastructure
+- No duplicate performance functionality
+- All performance follows existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-2", "TASK 2I", contract_2i))
+
+        contract_2j = f"""🎯 **AGENT-2 CONTRACT: TASK 2J - HEALTH SYSTEM FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize health system using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete health system finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document health system tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate health system uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Health system 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- System uses existing unified infrastructure
+- No duplicate health functionality
+- All health monitoring follows existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-2", "TASK 2J", contract_2j))
+
+        contract_2k = f"""🎯 **AGENT-2 CONTRACT: TASK 2K - API INTEGRATION FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize API integration using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete API integration finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document API integration tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate API integration uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- API integration 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Integration uses existing unified infrastructure
+- No duplicate API functionality
+- All API calls follow existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-2", "TASK 2K", contract_2k))
+
+        return contracts
