+++ src/task_distribution/agents/agent3.py
@@ -1,0 +1,122 @@
+from __future__ import annotations
+
+from ..models import TaskContract
+from ..standards import CODING_STANDARDS
+
+
+class Agent3Contracts:
+    """Generate contracts for Agent-3 tasks."""
+
+    @staticmethod
+    def build() -> list[TaskContract]:
+        contracts: list[TaskContract] = []
+
+        contract_3g = f"""🎯 **AGENT-3 CONTRACT: TASK 3G - TESTING INFRASTRUCTURE FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize testing infrastructure using existing unified systems
+
+**📋 DELIVERABLES**:
+1) Complete testing infrastructure finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document testing infrastructure tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate testing uses existing unified infrastructure
+
+**✅ EXPECTED RESULTS**:
+- Testing infrastructure 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Infrastructure uses existing unified systems
+- No duplicate testing functionality
+- All tests follow existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-3", "TASK 3G", contract_3g))
+
+        contract_3h = f"""🎯 **AGENT-3 CONTRACT: TASK 3H - REPORTING SYSTEMS FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize reporting systems using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete reporting systems finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document reporting systems tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate reporting uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Reporting systems 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Systems use existing unified infrastructure
+- No duplicate reporting functionality
+- All reporting follows existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-3", "TASK 3H", contract_3h))
+
+        contract_3i = f"""🎯 **AGENT-3 CONTRACT: TASK 3I - INTEGRATION TESTING FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize integration testing using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete integration testing finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document integration testing tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate integration testing uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Integration testing 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Testing uses existing unified infrastructure
+- No duplicate testing functionality
+- All integration tests follow existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-3", "TASK 3I", contract_3i))
+
+        contract_3j = f"""🎯 **AGENT-3 CONTRACT: TASK 3J - MODEL CONSOLIDATION FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize model consolidation using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete model consolidation finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document model consolidation tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate model consolidation uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Model consolidation 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- Consolidation uses existing unified infrastructure
+- No duplicate model functionality
+- All models follow existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-3", "TASK 3J", contract_3j))
+
+        return contracts
