+++ src/task_distribution/agents/agent4.py
@@ -1,0 +1,41 @@
+from __future__ import annotations
+
+from ..models import TaskContract
+from ..standards import CODING_STANDARDS
+
+
+class Agent4Contracts:
+    """Generate contracts for Agent-4 tasks."""
+
+    @staticmethod
+    def build() -> list[TaskContract]:
+        contracts: list[TaskContract] = []
+
+        contract_4j = f"""🎯 **AGENT-4 CONTRACT: TASK 4J - REPOSITORY SYSTEM FINALIZATION**
+
+{CODING_STANDARDS}
+
+**🎯 OBJECTIVE**: Finalize repository system using existing unified infrastructure
+
+**📋 DELIVERABLES**:
+1) Complete repository system finalization
+2) Create comprehensive devlog entry in `logs/` directory
+3) Document repository system tasks completed
+4) Report architecture compliance status (V2 standards)
+5) Validate repository system uses existing unified systems
+
+**✅ EXPECTED RESULTS**:
+- Repository system 100% complete and functional
+- Devlog created with detailed finalization documentation
+- Architecture compliance verified (≤200 LOC, SRP, OOP)
+- System uses existing unified infrastructure
+- No duplicate repository functionality
+- All repository operations follow existing patterns
+
+**⏰ TIMELINE**: 2-3 hours
+
+**🚀 START NOW. WE. ARE. SWARM.**
+"""
+        contracts.append(TaskContract("Agent-4", "TASK 4J", contract_4j))
+
+        return contracts
