+++ src/task_distribution/models.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class TaskContract:
+    """Representation of a single task contract."""
+    agent: str
+    task_id: str
+    content: str
